# car-reselling-backend/cmd/api/main.go
package main

import (
	"log"
	"net/http"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	"github.com/yourusername/car-reselling-backend/internal/auth"
	"github.com/yourusername/car-reselling-backend/internal/config"
	"github.com/yourusername/car-reselling-backend/internal/database"
	"github.com/yourusername/car-reselling-backend/internal/listing"
	"github.com/yourusername/car-reselling-backend/internal/models"

	_ "github.com/yourusername/car-reselling-backend/docs" // Swagger docs
)

// @title           Car Reselling Backend API
// @version         1.0
// @description     A high-performance backend for a second-hand car reselling marketplace
// @termsOfService  http://swagger.io/terms/

// @contact.name   API Support
// @contact.url    http://www.example.com/support
// @contact.email  support@example.com

// @license.name  MIT
// @license.url   http://www.apache.org/licenses/LICENSE-2.0.html

// @host      localhost:3000
// @BasePath  /

// @securityDefinitions.apikey BearerAuth
// @in header
// @name Authorization
// @description Type "Bearer" followed by a space and JWT token.

func main() {
	// Load configuration
	cfg, err := config.Load()
	if err != nil {
		log.Fatalf("Failed to load configuration: %v", err)
	}

	// Initialize database connections
	if err := database.InitPostgres(cfg.DatabaseURL); err != nil {
		log.Fatalf("Failed to initialize PostgreSQL: %v", err)
	}
	defer database.Close()

	if err := database.InitRedis(cfg.RedisURL); err != nil {
		log.Fatalf("Failed to initialize Redis: %v", err)
	}
	defer database.CloseRedis()

	// Check if tables exist before running auto-migrations
	// This avoids constraint name conflicts when tables are created via SQL migrations
	var tableExists bool
	database.DB.Raw("SELECT EXISTS (SELECT FROM information_schema.tables WHERE table_schema = 'public' AND table_name = 'users')").Scan(&tableExists)

	if !tableExists {
		// Tables don't exist, run auto-migrations
		if err := database.DB.AutoMigrate(&models.User{}, &models.VerificationCode{}); err != nil {
			log.Fatalf("Failed to run migrations: %v", err)
		}
		log.Println("Database tables created successfully")
	} else {
		// Tables exist, skip auto-migration to avoid constraint conflicts
		log.Println("Database tables already exist, skipping auto-migration")
		// Still try to sync schema for new fields (non-destructive)
		if err := database.DB.AutoMigrate(&models.User{}, &models.VerificationCode{}); err != nil {
			// Log but don't fail - constraint name mismatches are OK
			log.Printf("Note: Schema sync encountered minor issues (this is normal): %v", err)
		}
	}

	// Setup Gin router
	if cfg.Environment == "production" {
		gin.SetMode(gin.ReleaseMode)
	}

	r := gin.Default()

	// Middleware
	r.Use(gin.Recovery())
	r.Use(gin.Logger())

	// Rate limiting (exclude Swagger UI and health check)
	r.Use(func(c *gin.Context) {
		path := c.Request.URL.Path
		// Skip rate limiting for Swagger UI and health check
		if path == "/health" ||
			path == "/swagger" ||
			path == "/swagger/" ||
			(len(path) > 9 && path[:9] == "/swagger/") {
			c.Next()
			return
		}
		// Apply rate limiting for other routes
		auth.RateLimitMiddleware()(c)
	})

	// CORS middleware (for development)
	r.Use(func(c *gin.Context) {
		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, DELETE, PATCH")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	})

	// Swagger UI
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	// Health check endpoint
	// @Summary Health check
	// @Description Check if the API and database are healthy
	// @Tags health
	// @Produce json
	// @Success 200 {object} map[string]string
	// @Failure 503 {object} map[string]string
	// @Router /health [get]
	r.GET("/health", func(c *gin.Context) {
		if err := database.HealthCheck(); err != nil {
			c.JSON(http.StatusServiceUnavailable, gin.H{
				"status": "unhealthy",
				"error":  err.Error(),
			})
			return
		}
		c.JSON(http.StatusOK, gin.H{
			"status": "healthy",
		})
	})

	// Initialize auth components
	authRepo := auth.NewRepository()
	authService := auth.NewService(authRepo, cfg)
	authHandler := auth.NewHandler(authService)

	// Auth routes
	authGroup := r.Group("/api/auth")
	{
		authGroup.POST("/register", authHandler.Register)
		authGroup.POST("/login", authHandler.Login)
		authGroup.POST("/send-otp", authHandler.SendOTP)
		authGroup.POST("/verify-otp", authHandler.VerifyOTP)
		authGroup.POST("/refresh", authHandler.RefreshToken)

		// Protected routes
		protected := authGroup.Group("")
		protected.Use(auth.AuthMiddleware(cfg))
		{
			protected.POST("/logout", authHandler.Logout)
			protected.GET("/me", authHandler.GetCurrentUser)
		}
	}

	// Initialize listing components
	listingRepo := listing.NewRepository(database.DB)
	storageService, err := listing.NewStorageService(cfg.AWSAccessKey, cfg.AWSSecretKey, cfg.AWSRegion, cfg.AWSBucketName)
	if err != nil {
		log.Printf("Failed to initialize storage service: %v", err)
	}
	listingService := listing.NewService(listingRepo, storageService, database.RedisClient)
	listingHandler := listing.NewHandler(listingService)

	// Listing routes
	// Listing routes
	api := r.Group("/api")
	{
		cars := api.Group("/cars")

		// Public listing routes
		cars.GET("", listingHandler.ListListings)
		cars.GET("/:id", listingHandler.GetListing)

		// Protected listing routes
		protected := cars.Group("")
		protected.Use(auth.AuthMiddleware(cfg))
		{
			protected.POST("", listingHandler.CreateListing)
			protected.PUT("/:id", listingHandler.UpdateListing)
			protected.DELETE("/:id", listingHandler.DeleteListing)

			// Custom endpoints (careful with path conflicts, but these are distinct enough)
			// :id matches UUIDs usually, so "favorites" and "my-listings" might conflict if :id is catch-all.
			// Gin matches static paths first, so /cars/favorites should work before /cars/:id if :id is not regex restricted.
			// Best practice: define static paths BEFORE param paths in definition order if possible, though Gin handles it.

			// Actually, putting these inside the group might be tricky if :id is defined on the same level.
			// Let's check: /api/cars/:id vs /api/cars/my-listings
			// Gin router handles static vs param, but good to be explicit.
		}

		// Specific protected routes that shouldn't clash with :id
		// To be safe, define them before :id if they were in the same group context,
		// or relies on Gin's priority. Gin prioritizes static matches.

		protectedListings := api.Group("/cars")
		protectedListings.Use(auth.AuthMiddleware(cfg))

		protectedListings.GET("/my-listings", listingHandler.GetMyListings)
		protectedListings.GET("/favorites", listingHandler.GetFavorites)
		protectedListings.POST("/:id/favorite", listingHandler.ToggleFavorite)
	}

	// Start server
	serverAddr := ":" + cfg.ServerPort
	log.Printf("Server starting on %s", serverAddr)
	if err := http.ListenAndServe(serverAddr, r); err != nil {
		log.Fatalf("Failed to start server: %v", err)
	}
}